refactor, move export/import to procedure
