+++ minigu/main/Cargo.toml
@@ -3,9 +3,14 @@ edition.workspace = true
 name = "minigu"
 version = "0.1.0"
 
+[dev-dependencies]
+tempfile = "3.20.0"
+walkdir = "2.5.0"
+
 [dependencies]
 anyhow = { workspace = true }
 arrow = { workspace = true }
+csv = { workspace = true }
 gql-parser = { workspace = true }
 itertools = { workspace = true }
 miette = { workspace = true }
@@ -18,6 +23,8 @@ minigu-ir = { workspace = true }
 minigu-planner = { workspace = true }
 minigu-storage = { workspace = true }
 rayon = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
 thiserror = { workspace = true }
 
 [lints]
