+++ minigu/main/src/procedures/export_import/mod.rs
@@ -21,25 +21,21 @@
 //! <eid>,<src‑vid>,<dst‑vid>,<prop‑1>,<prop‑2>, ...
 //! ```
 
-use core::error::Error;
-use std::collections::{BTreeMap, HashMap, HashSet};
-use std::fs::File;
-use std::path::Path;
+use std::collections::{HashMap, HashSet};
+use std::error::Error;
 use std::str::FromStr;
 use std::sync::Arc;
 
-use csv::{ReaderBuilder, Writer, WriterBuilder};
 use minigu_catalog::label_set::LabelSet;
 use minigu_catalog::property::Property;
 use minigu_catalog::provider::GraphTypeProvider;
-use minigu_common::data_type::LogicalType;
-use minigu_common::types::{EdgeId, LabelId, VertexId};
-use minigu_common::value::ScalarValue;
+use minigu_common::types::LabelId;
 use serde::{Deserialize, Serialize};
 
-use crate::common::{Edge, PropertyRecord, Vertex};
-use crate::tp::{MemoryGraph, TransactionHandle};
+pub mod export;
+pub mod import;
 
+type Result<T> = std::result::Result<T, Box<dyn Error + Send + Sync + 'static>>;
 type RecordType = Vec<String>;
 
 /// Cached lookup information derived from `GraphTypeProvider`.
@@ -112,18 +108,18 @@ impl FileSpec {
 
 /// Common metadata for a vertex or edge collection.
 #[derive(Deserialize, Serialize, Debug)]
-pub struct VertexSpec {
+struct VertexSpec {
     label: String,
     file: FileSpec,
     properties: Vec<Property>,
 }
 
 impl VertexSpec {
-    pub fn label_name(&self) -> &String {
+    fn label_name(&self) -> &String {
         &self.label
     }
 
-    pub fn properties(&self) -> &Vec<Property> {
+    fn properties(&self) -> &Vec<Property> {
         &self.properties
     }
 
@@ -162,26 +158,26 @@ impl EdgeSpec {
         }
     }
 
-    pub fn src_label(&self) -> &String {
+    fn src_label(&self) -> &String {
         &self.src_label
     }
 
-    pub fn dst_label(&self) -> &String {
+    fn dst_label(&self) -> &String {
         &self.dst_label
     }
 
-    pub fn label_name(&self) -> &String {
+    fn label_name(&self) -> &String {
         &self.label
     }
 
-    pub fn properties(&self) -> &Vec<Property> {
+    fn properties(&self) -> &Vec<Property> {
         &self.properties
     }
 }
 
 // Top-level manifest written to disk.
 #[derive(Deserialize, Serialize, Default, Debug)]
-pub struct Manifest {
+struct Manifest {
     vertices: Vec<VertexSpec>,
     edges: Vec<EdgeSpec>,
 }
@@ -267,294 +263,114 @@ impl Manifest {
 impl FromStr for Manifest {
     type Err = Box<dyn Error + Send + Sync + 'static>;
 
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
+    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
         Ok(serde_json::from_str(s)?)
-    }
-}
-
-/// Convert a *string* coming from CSV into an owned [`ScalarValue`] according
-/// to a given property definition.
-fn property_to_scalar_value(property: &Property, value: &str) -> ScalarValue {
-    match property.logical_type() {
-        LogicalType::Int8 => ScalarValue::Int8(value.parse().ok()),
-        LogicalType::Int16 => ScalarValue::Int16(value.parse().ok()),
-        LogicalType::Int32 => ScalarValue::Int32(value.parse().ok()),
-        LogicalType::Int64 => ScalarValue::Int64(value.parse().ok()),
-        LogicalType::UInt8 => ScalarValue::UInt8(value.parse().ok()),
-        LogicalType::UInt16 => ScalarValue::UInt16(value.parse().ok()),
-        LogicalType::UInt32 => ScalarValue::UInt32(value.parse().ok()),
-        LogicalType::UInt64 => ScalarValue::UInt64(value.parse().ok()),
-        LogicalType::Boolean => ScalarValue::Boolean(value.parse().ok()),
-        LogicalType::Float32 => ScalarValue::Float32(value.parse().ok()),
-        LogicalType::Float64 => ScalarValue::Float64(value.parse().ok()),
-        LogicalType::String => ScalarValue::String(Some(value.to_string())),
-        _ => todo!(),
-    }
-}
-
-/// Convert a [`ScalarValue`] back into a *CSV‑ready* string. `NULL` becomes an
-/// empty string.
-fn scalar_value_to_string(scalar_value: &ScalarValue) -> String {
-    match scalar_value {
-        ScalarValue::Int8(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::Int16(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::Int32(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::Int64(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::UInt8(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::UInt16(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::UInt32(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::UInt64(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::Boolean(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::Float32(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::Float64(value) => value.map_or(String::new(), |inner| inner.to_string()),
-        ScalarValue::String(value) => value.clone().unwrap_or_default(),
-        _ => todo!(),
-    }
-}
-
-#[derive(Debug)]
-struct VerticesBuilder {
-    records: HashMap<LabelId, BTreeMap<VertexId, RecordType>>,
-    writers: HashMap<LabelId, Writer<File>>,
-}
-
-impl VerticesBuilder {
-    fn new<P: AsRef<Path>>(dir: P, map: &HashMap<LabelId, String>) -> Self {
-        let writers = map
-            .iter()
-            .map(|(id, label)| {
-                let filename = format!("{}.csv", label);
-                let path = dir.as_ref().join(filename);
-
-                (*id, WriterBuilder::new().from_path(path).unwrap())
-            })
-            .collect();
-        Self {
-            records: HashMap::new(),
-            writers,
-        }
-    }
-
-    fn add_vertex(&mut self, v: &Vertex) {
-        let record = std::iter::once(v.vid().to_string())
-            .chain(v.properties().iter().map(scalar_value_to_string))
-            .collect::<Vec<_>>();
-
-        self.records
-            .entry(v.label_id)
-            .or_default()
-            .insert(v.vid(), record);
-    }
-
-    fn dump(&mut self) {
-        for (label_id, records) in self.records.iter() {
-            let w = self.writers.get_mut(label_id).unwrap();
-
-            for (_, record) in records.iter() {
-                w.write_record(record).unwrap();
-            }
-        }
-    }
-}
-
-#[derive(Debug)]
-struct EdgesBuilder {
-    records: HashMap<LabelId, BTreeMap<EdgeId, RecordType>>,
-    writers: HashMap<LabelId, Writer<File>>,
-}
-
-impl EdgesBuilder {
-    fn new<P: AsRef<Path>>(dir: P, map: &HashMap<LabelId, String>) -> Self {
-        let writers = map
-            .iter()
-            .map(|(id, label)| {
-                let filename = format!("{}.csv", label);
-                let path = dir.as_ref().join(filename);
-
-                (*id, WriterBuilder::new().from_path(path).unwrap())
-            })
-            .collect();
-
-        Self {
-            records: HashMap::new(),
-            writers,
-        }
-    }
-
-    pub fn add_edge(&mut self, e: &Edge) {
-        let prefix = [
-            e.eid().to_string(),
-            e.src_id().to_string(),
-            e.dst_id().to_string(),
-        ];
-        let record = prefix
-            .into_iter()
-            .chain(e.properties().iter().map(scalar_value_to_string))
-            .collect::<Vec<_>>();
-
-        self.records
-            .entry(e.label_id)
-            .or_default()
-            .insert(e.eid(), record);
-    }
-
-    pub fn dump(&mut self) {
-        for (label_id, records) in self.records.iter() {
-            let w = self.writers.get_mut(label_id).unwrap();
-
-            for (_, record) in records.iter() {
-                w.write_record(record).unwrap();
-            }
-        }
-    }
-}
-
-impl MemoryGraph {
-    pub fn import_graph<P: AsRef<Path>>(
-        &self,
-        txn: &TransactionHandle,
-        manifest_path: P,
-        graph_type: Arc<dyn GraphTypeProvider>,
-    ) {
-        // 1. Reading manifest
-        let manifest_str = std::fs::read_to_string(manifest_path.as_ref()).unwrap();
-        let manifest: Manifest = serde_json::from_str(&manifest_str).unwrap();
-
-        // 2. Vertices
-        let mut vid = 1;
-        for vertex_spec in manifest.vertices.iter() {
-            let mut rdr = ReaderBuilder::new()
-                .has_headers(false)
-                .from_path(
-                    manifest_path
-                        .as_ref()
-                        .parent()
-                        .unwrap()
-                        .join(&vertex_spec.file.path),
-                )
-                .unwrap();
-
-            let label_id = graph_type
-                .get_label_id(&vertex_spec.label)
-                .unwrap()
-                .unwrap();
-
-            for record in rdr.records() {
-                let label_set = LabelSet::from_iter(vec![label_id]);
-                let props_schema = graph_type
-                    .get_vertex_type(&label_set)
-                    .unwrap()
-                    .unwrap()
-                    .properties();
-
-                let record = record.unwrap();
-
-                assert_eq!(props_schema.len() + 1, record.len());
-
-                let props = props_schema
-                    .iter()
-                    .zip(record.iter().skip(1)) // skip vid
-                    .map(|((_, p), value)| property_to_scalar_value(p, value))
-                    .collect();
-                let vertex = Vertex::new(vid, label_id, PropertyRecord::new(props));
-
-                self.create_vertex(txn, vertex).unwrap();
-                vid += 1;
-            }
-        }
-
-        // 3. Edges
-        let mut eid = 1;
-        for edge_spec in manifest.edges.iter() {
-            let label_id = graph_type.get_label_id(&edge_spec.label).unwrap().unwrap();
-
-            let mut rdr = ReaderBuilder::new()
-                .has_headers(false)
-                .from_path(
-                    manifest_path
-                        .as_ref()
-                        .parent()
-                        .unwrap()
-                        .join(&edge_spec.file.path),
-                )
-                .unwrap();
-
-            for record in rdr.records() {
-                let label_set = LabelSet::from_iter(vec![label_id]);
-
-                let props = graph_type
-                    .get_edge_type(&label_set)
-                    .unwrap()
-                    .unwrap()
-                    .properties();
-
-                let record = record.unwrap();
-
-                assert_eq!(record.len() - 3, props.len());
-                let src_id = record.get(1).unwrap().parse().unwrap();
-                let dst_id = record.get(2).unwrap().parse().unwrap();
-
-                let props = props
-                    .iter()
-                    .zip(record.iter().skip(3)) // skip eid, src and dst
-                    .map(|((_, p), value)| property_to_scalar_value(p, value))
-                    .collect();
-
-                let edge = Edge::new(eid, src_id, dst_id, label_id, PropertyRecord::new(props));
-                self.create_edge(txn, edge).unwrap();
-                eid += 1;
-            }
-        }
-    }
-
-    pub fn export_graph<P: AsRef<Path>>(
-        &self,
-        txn: &TransactionHandle,
-        dir: P,
-        manifest_rel_path: P, // relative path
-        graph_type: Arc<dyn GraphTypeProvider>,
-    ) {
-        // 1. Prepare output paths
-        let dir = dir.as_ref();
-        std::fs::create_dir_all(dir).unwrap();
-
-        let metadata = SchemaMetadata::from_schema(Arc::clone(&graph_type));
-
-        let mut vertice_builder = VerticesBuilder::new(dir, &metadata.label_map);
-        let mut edges_builder = EdgesBuilder::new(dir, &metadata.label_map);
-
-        // 2. Dump vertices
-        for v in txn.iter_vertices() {
-            vertice_builder.add_vertex(&v.unwrap());
-        }
-        vertice_builder.dump();
-
-        // 3. Dump edge
-        for e in txn.iter_edges() {
-            edges_builder.add_edge(&e.unwrap());
-        }
-        edges_builder.dump();
-
-        // 4. Dump manifest
-        let manifest = Manifest::from_schema(metadata);
-        std::fs::write(
-            dir.join(manifest_rel_path),
-            serde_json::to_string(&manifest).unwrap(),
-        )
-        .unwrap();
     }
 }
 
 #[cfg(test)]
 mod test {
+    use std::collections::BTreeMap;
+    use std::path::Path;
+
     use minigu_catalog::memory::graph_type::{
         MemoryEdgeTypeCatalog, MemoryGraphTypeCatalog, MemoryVertexTypeCatalog,
     };
+    use minigu_common::data_type::LogicalType;
+    use minigu_common::types::{EdgeId, VertexId};
+    use minigu_common::value::ScalarValue;
+    use minigu_storage::common::{Edge, PropertyRecord, Vertex};
+    use minigu_storage::tp::memory_graph::{self};
+    use minigu_storage::tp::{IsolationLevel, MemoryGraph};
     use walkdir::WalkDir;
 
     use super::*;
-    use crate::tp::IsolationLevel;
-    use crate::tp::memory_graph::tests::*;
+    use crate::procedures::export_import::export::export;
+    use crate::procedures::export_import::import::import;
+
+    const PERSON: LabelId = LabelId::new(1).unwrap();
+    const FRIEND: LabelId = LabelId::new(2).unwrap();
+    const FOLLOW: LabelId = LabelId::new(3).unwrap();
+
+    fn create_vertex(vid: VertexId, label_id: LabelId, properties: Vec<ScalarValue>) -> Vertex {
+        Vertex::new(vid, label_id, PropertyRecord::new(properties))
+    }
+
+    fn create_edge(
+        eid: EdgeId,
+        src_id: VertexId,
+        dst_id: VertexId,
+        label_id: LabelId,
+        properties: Vec<ScalarValue>,
+    ) -> Edge {
+        Edge::new(
+            eid,
+            src_id,
+            dst_id,
+            label_id,
+            PropertyRecord::new(properties),
+        )
+    }
+
+    fn mock_graph() -> Arc<MemoryGraph> {
+        let graph = MemoryGraph::new();
+
+        let txn = graph.begin_transaction(IsolationLevel::Serializable);
+
+        let alice = create_vertex(1, PERSON, vec![
+            ScalarValue::String(Some("Alice".to_string())),
+            ScalarValue::Int32(Some(25)),
+        ]);
+
+        let bob = create_vertex(2, PERSON, vec![
+            ScalarValue::String(Some("Bob".to_string())),
+            ScalarValue::Int32(Some(28)),
+        ]);
+
+        let carol = create_vertex(3, PERSON, vec![
+            ScalarValue::String(Some("Carol".to_string())),
+            ScalarValue::Int32(Some(24)),
+        ]);
+
+        let david = create_vertex(4, PERSON, vec![
+            ScalarValue::String(Some("David".to_string())),
+            ScalarValue::Int32(Some(27)),
+        ]);
+
+        // Add vertices to the graph
+        graph.create_vertex(&txn, alice).unwrap();
+        graph.create_vertex(&txn, bob).unwrap();
+        graph.create_vertex(&txn, carol).unwrap();
+        graph.create_vertex(&txn, david).unwrap();
+
+        // Create friend edges
+        let friend1 = create_edge(1, 1, 2, FRIEND, vec![ScalarValue::String(Some(
+            "2020-01-01".to_string(),
+        ))]);
+
+        let friend2 = create_edge(2, 2, 3, FRIEND, vec![ScalarValue::String(Some(
+            "2021-03-15".to_string(),
+        ))]);
+
+        // Create follow edges
+        let follow1 = create_edge(3, 1, 3, FOLLOW, vec![ScalarValue::String(Some(
+            "2022-06-01".to_string(),
+        ))]);
+
+        let follow2 = create_edge(4, 4, 1, FOLLOW, vec![ScalarValue::String(Some(
+            "2022-07-15".to_string(),
+        ))]);
+
+        // Add edges to the graph
+        graph.create_edge(&txn, friend1).unwrap();
+        graph.create_edge(&txn, friend2).unwrap();
+        graph.create_edge(&txn, follow1).unwrap();
+        graph.create_edge(&txn, follow2).unwrap();
+
+        txn.commit().unwrap();
+
+        graph
+    }
 
     fn mock_graph_type() -> MemoryGraphTypeCatalog {
         let mut graph_type = MemoryGraphTypeCatalog::new();
@@ -678,30 +494,21 @@ mod test {
 
         let graph_type: Arc<dyn GraphTypeProvider> = Arc::new(mock_graph_type());
         {
-            let (graph, _cleaner) = mock_graph();
-            let txn = graph.begin_transaction(IsolationLevel::Serializable);
+            let graph = mock_graph();
 
-            graph.export_graph(
-                &txn,
+            export(
+                graph,
                 export_dir1,
                 manifest_rel_path.as_ref(),
                 Arc::clone(&graph_type),
             );
-            txn.commit().unwrap();
         }
 
         {
-            let (graph, _) = mock_empty_graph();
-            let txn = graph.begin_transaction(IsolationLevel::Serializable);
+            let manifest_path = export_dir1.join(manifest_rel_path);
+            let (graph, graph_type) = import(manifest_path).unwrap();
 
-            graph.import_graph(
-                &txn,
-                export_dir1.join(manifest_rel_path),
-                Arc::clone(&graph_type),
-            );
-
-            graph.export_graph(&txn, export_dir2, manifest_rel_path.as_ref(), graph_type);
-            txn.commit().unwrap();
+            export(graph, export_dir2, manifest_rel_path.as_ref(), graph_type);
         }
 
         assert!(export_dirs_equal_semantically(export_dir1, export_dir2))
