+++ minigu/main/src/procedures/mod.rs
@@ -1,7 +1,6 @@
 mod create_test_graph;
 mod echo;
-mod export;
-mod import;
+mod export_import;
 mod show_procedures;
 
 use minigu_context::procedure::Procedure;
@@ -17,7 +16,13 @@ pub fn build_predefined_procedures() -> Vec<(String, Procedure)> {
             "create_test_graph".to_string(),
             create_test_graph::build_procedure(),
         ),
-        ("import".to_string(), import::build_procedure()),
-        ("export".to_string(), export::build_procedure()),
+        (
+            "import".to_string(),
+            export_import::import::build_procedure(),
+        ),
+        (
+            "export".to_string(),
+            export_import::export::build_procedure(),
+        ),
     ]
 }
